requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.5
certifi==2024.12.14
charset-normalizer==3.4.1
cloudinary==1.44.1
defusedxml==0.7.1
Django==4.2.7
django-cloudinary-storage==0.3.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.19.1
gunicorn==23.0.0
httpie==3.2.4
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.3
nodejs==0.1.1
optional-django==0.1.0
packaging==25.0
pillow==11.0.0
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-whois==0.9.5
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.0.0
setuptools==80.3.1
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tldextract==5.3.0
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
whitenoise==6.9.0
